Introduce ADC library and test
* ADC library implemented and tested
* Minor changes in UART implementation

+++ adc/test/mocks/adc_mock.cpp
@@ -7,7 +7,7 @@ extern "C" void adc_init(void){
 	return Fixture::_adc -> adc_init();
 }
 
-extern "C" uint8_t read_angle(void){
+extern "C" uint16_t read_angle(void){
 	return Fixture::_adc -> read_angle();
 }
